move proxyreader
